server/requirements.txt
# Web Frameworks
Flask==3.0.0
fastapi==0.95.0

# Web Servers & ASGI
uvicorn==0.21.1

# Database
pymongo==4.6.1
Flask-PyMongo==2.3.0

# Security & Authentication
Flask-JWT-Extended==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Configuration
python-dotenv==1.0.1

# Data Validation & Parsing
pydantic[email]==1.10.7
python-multipart==0.0.6

# CORS
Flask-CORS==4.0.0

# AI Integration
openai==0.27.4

# Optional Tools
APScheduler==3.10.1

# Testing & Development (optional, comment out if not needed)
pytest==8.0.2
pytest-flask==1.3.0

# Linting & Formatting (optional, comment out if not needed)
black==24.2.0
flake8==7.0.0
isort==5.10.3
mypy==0.910